+++ crates/core/src/catalog.rs
@@ -55,6 +55,13 @@ pub struct Catalog {
 }
 
 #[serde(rename_all = "camelCase")]
+#[derive(Debug, Clone, Deserialize, Eq, PartialEq, Ord, PartialOrd)]
+pub struct GameVersion {
+    pub game_version: String,
+    pub flavor: Flavor,
+}
+
+#[serde(rename_all = "camelCase")]
 #[derive(Debug, Clone, Deserialize)]
 pub struct CatalogAddon {
     pub id: u32,
@@ -66,7 +73,9 @@ pub struct CatalogAddon {
     pub summary: String,
     pub number_of_downloads: u64,
     pub source: Source,
+    #[deprecated(since = "0.4.4", note = "Please use game_versions instead")]
     pub flavors: Vec<Flavor>,
+    pub game_versions: Vec<GameVersion>,
 }
 
 mod date_parser {